Stop calling IFormatterResolver to serialize string/byte[]
Consider string and byte[] to be "optimized" like the other primitive types now that the OldSpec behavior is handled by the MessagePackWriter itself rather than an OldspecResolver.

In PerfNetFramework, this drops the cost of  clr!JIT_VirtualFunctionPointer from ~32% to just under 18%. The remaining time is spent calling a resolver for an enum.

Fixes most of #48