Add public API tracking files
This makes it much easier to avoid accidental public API breaking changes, particularly since a good portion of the public API is generated with .tt files.

If you're not already familiar with this mechanism, it adds a Roslyn analyzer that produces build warnings when API is added, changed or removed. A code fix is offered that updates the API tracking files to match code in order to eliminate the warning. This action requires acknowledging of the API change and changes an SCC-tracked file so that it is also highlighted in a code review.

Right now since v2.0 includes many API breaking changes, all APIs are in the "unshipped" file. Once 2.0 is stable, we will bulk move all these APIs to the "shipped" file, to signify the severity of changes to shipped APIs is much higher.